More refactoring changes.
