+++ lib/qrsChangeOwner.js
@@ -1,7 +1,7 @@
 var Promise = require('bluebird')
 var config = require('../config/config');
 var winston = require('winston');
-var qrsInteract = require('./qrsinteractions');
+var qrsInteract = require('./qrsInstance');
 var qrsNotify = require('./qrsNotify');
 
 
@@ -20,9 +20,8 @@ var qrsChangeOwner = {
         //We are doing this because we want to make sure master library items are visible on unpublished apps for developers creating apps.
         return new Promise(function(resolve, reject) {
             logger.debug('running the changeAgent for the Governed Metrics Service', {module:'qrsChangeOwner',method:'changeAgent'});
-            var path = "https://" + config.hostname + ":" + config.qrsPort + "/qrs/app/object";
-            path += "?xrfkey=ABCDEFG123456789";
-            path += "&filter=owner.userId eq '" + config.repoAccountUserId + "' and owner.userDirectory eq '";
+            var path = "/app/object";
+            path += "?filter=owner.userId eq '" + config.repoAccountUserId + "' and owner.userDirectory eq '";
             path += config.repoAccountUserDirectory + "' and (objectType eq 'dimension' or objectType eq 'measure')";
             return qrsInteract.get(path)
             .then(function(appObjects) {
@@ -57,8 +56,7 @@ var qrsChangeOwner = {
                 })
                 .then(function(body) {
                     //logger.debug('Create selection on app objects')
-                    var postPath = "https://" + config.hostname + ":" + config.qrsPort + "/qrs/selection";
-                    postPath += "?xrfkey=ABCDEFG123456789";
+                    var postPath = "/selection";
                     //logger.debug('qrsInteract.post on path:' + postPath, {module:'qrsChangeOwner',method:'processAppObjects'});
                     return qrsInteract.post(postPath, body, 'json');
                 })
@@ -116,8 +114,7 @@ var qrsChangeOwner = {
                 .then(function(body) {
                     logger.debug('Body of objects for selection:' + JSON.stringify(body),{module:'qrsChangeOwner'});          
 
-                    var postPath = "https://" + config.hostname + ":" + config.qrsPort + "/qrs/selection";
-                    postPath += "?xrfkey=ABCDEFG123456789";
+                    var postPath = "/selection";
                     logger.debug("qrsChangeOwner::postPath::" + postPath, {module: 'qrsChangeOwner'});
                     qrsInteract.post(postPath, body, 'json')
                         .then(function(result) {
@@ -134,8 +131,7 @@ var qrsChangeOwner = {
                                 }]
                             };
                             //logger.debug('qrsChangeOwner::Body for changing owner on app.object. ' +  JSON.stringify(body), {module: 'qrsChangeOwner'});
-                            var putPath = "https://" + config.hostname + ":" + config.qrsPort + "/qrs/selection/" + result.id + "/app/object/synthetic";
-                            putPath += "?xrfkey=ABCDEFG123456789"
+                            var putPath = "/selection/" + result.id + "/app/object/synthetic";
                                 //logger.debug('qrsChangeOwner::PutPath::' + putPath, {module: 'qrsChangeOwner'});
                             qrsInteract.put(putPath, body)
                                 .then(function(sCode) {
@@ -143,8 +139,7 @@ var qrsChangeOwner = {
                                     logger.info(sCode + ' received during Put operation for changing ownership',{module: 'qrsChangeOwner'});
                                     logger.info('delete this selection:' + x.id, {module: 'qrsChangeOwner'});
                                     //logger.debug('qrsChangeOwner::Changed ownership to user ' + ownerId, {module: 'qrsChangeOwner'});
-                                    var deletePath = "https://" + config.hostname + ":" + config.qrsPort + "/qrs/selection/" + x.id;
-                                    deletePath += "/?xrfkey=ABCDEFG123456789";
+                                    var deletePath = "/selection/" + x.id + "/";
                                     //logger.debug(x.id, {module: 'qrsChangeOwner'});
                                     qrsInteract.delete(deletePath)
                                         .then(function(result) {
@@ -198,9 +193,8 @@ function segregateAppObjects(appObjects) {
             {
                 appCount++;
             //logger.debug(appCount, ' of ', array.length,  {module: 'qrsChangeOwner', method: 'segregateAppObjects'});
-                var path = "https://" + config.hostname + ":" + config.qrsPort + "/qrs/app/full";
-                path += "?xrfkey=ABCDEFG123456789";
-                path += "&filter=id eq " + item;
+                var path = "/app/full";
+                path += "?filter=id eq " + item;
                 return qrsInteract.get(path)
                 .then(function(result)
                 {
@@ -255,9 +249,8 @@ function segregateAppObjects(appObjects) {
 
 function getOwnerId(appId) {
     return new Promise(function(resolve,reject) {
-        var path = "https://" + config.hostname + ":" + config.qrsPort + "/qrs/app/full";
-        path += "?xrfkey=ABCDEFG123456789";
-        path += "&filter=id eq " + appId;
+        var path = "/app/full";
+        path += "?filter=id eq " + appId;
         qrsInteract.get(path)
             .then(function(result) {
                 //logger.debug('Owner is: ', result[0].owner.id, {module: 'qrsChangeOwner', method: 'segregateAppObjects'});
