+++ lib/qrsNotify.js
@@ -1,7 +1,7 @@
 var Promise = require('bluebird')
 var config = require('../config/config');
 var winston = require('winston');
-var qrsInteract = require('./qrsinteractions');
+var qrsInteract = require('./qrsInstance');
 
 //set up logging
 var logger = new (winston.Logger)({
@@ -19,9 +19,8 @@ var qrsNotify = {
         //to the repository.
         return new Promise(function(resolve, reject)
         {
-            var path = "https://" + config.hostname + ":" + config.qrsPort + "/qrs/notification";
-            path += "?xrfkey=ABCDEFG123456789";
-            path += "&name=appobject&changetype=Add";
+            var path = "/notification";
+            path += "?name=appobject&changetype=Add";
             path += "&filter=owner.userId eq '" + config.repoAccountUserId + "' and owner.userDirectory eq '" + config.repoAccountUserDirectory + "'";
             var body = "http://" + config.hostname + ":" + config.port + "/masterlib/notifyme";
             logger.debug('Creating notification service entry.', {module: 'qrsNotifyCreation', method: 'setNotification'});
@@ -42,9 +41,8 @@ var qrsNotify = {
     {
         return new Promise(function(resolve, reject)
         {
-            var path = "https://" + config.hostname + ":" + config.qrsPort + "/qrs/notification";
-            path += "?xrfkey=ABCDEFG123456789";
-            path += "&handle=" + notifyHandle;
+            var path = "/notification";
+            path += "?handle=" + notifyHandle;
             logger.debug('deleting notification service entry: ' + notifyHandle, {module: 'qrsNotifyCreation', method: 'delNotification'});
             qrsInteract.delete(path)
             .then(function(result)
