+++ lib/reloadmetrics.js
@@ -1,7 +1,7 @@
 var Promise = require('bluebird');
 var config = require('../config/config');
 var winston = require('winston');
-var qrsInteract = require('./qrsinteractions');
+var qrsInteract = require('./qrsInstance');
 var appDataSegment = require('./getAppDataSegment');
 
 //set up logging
@@ -26,8 +26,8 @@ var reloadMetrics = {
 			{
 				currentContentHash = result[0].contentHash;
 				logger.info('reloadMetrics::Starting Task::' + taskName, {module: 'reloadmetrics'});
-				var path = "https://" + config.hostname + ":" + config.qrsPort + "/qrs/task/start/synchronous";
-				path += "?xrfkey=ABCDEFG123456789&name=" + taskName;
+				var path = "/task/start/synchronous";
+				path += "?name=" + taskName;
 				logger.debug('reloadMetrics::PATH::' + path, {module: 'reloadmetrics'});
 				qrsInteract.post(path)
 				.then(function(result)
